requirement.txt
discord-py-slash-command
discord-ext-bot
discord.py
discord-ui
json2json
requests